Develop

+++ src/main/java/com/dev/gestorNF/business/VendedorService.java
@@ -5,6 +5,7 @@ import com.dev.gestorNF.business.dto.out.VendedorDTOResponse;
 import com.dev.gestorNF.business.mapper.VendedorConverter;
 import com.dev.gestorNF.infrastructure.entity.out.UsuarioEntity;
 import com.dev.gestorNF.infrastructure.entity.out.VendedorEntity;
+import com.dev.gestorNF.infrastructure.repository.NotaFiscalRepository;
 import com.dev.gestorNF.infrastructure.repository.UsuarioRepository;
 import com.dev.gestorNF.infrastructure.repository.VendedorRepository;
 import com.dev.gestorNF.infrastructure.security.JwtUtil;
@@ -21,12 +22,15 @@ public class VendedorService {
     private final JwtUtil jwtUtil;
     private final UsuarioRepository usuarioRepository;
     private final VendedorRepository vendedorRepository;
+    private final NotaFiscalRepository notaFiscalRepository;
 
     public VendedorDTOResponse cadastroVendedor(String token, VendedorDTORequest vendedorDTORequest) {
 
         String email = jwtUtil.extrairEmailToken(token.substring(7));
         UsuarioEntity usuarioEntity = usuarioRepository.findByEmail(email)
                 .orElseThrow(() -> new RuntimeException("Email não encontrado " + email));
+
+        validarComissao(usuarioEntity, vendedorDTORequest.getComissao());
 
         VendedorEntity vendedorEntity =
                 vendedorConverter.paraVendedorEntity(vendedorDTORequest);
@@ -39,15 +43,31 @@ public class VendedorService {
         return vendedorConverter.paraVendedorDTOResponse(vendedorSalvo);
     }
 
-    public void deletaVendedor(String token ,Long idVendedor) {
+    public void deletaVendedor(String token, Long idVendedor) {
 
         String email = jwtUtil.extrairEmailToken(token.substring(7));
 
         UsuarioEntity usuarioEntity = usuarioRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("Email não encontrado " + email));
+                .orElseThrow(() ->
+                        new RuntimeException("Email não encontrado " + email));
 
-        VendedorEntity vendedorEntity = vendedorRepository.findByIdVendedorAndUsuarioId(idVendedor,usuarioEntity.getId())
-                .orElseThrow(() -> new RuntimeException("Vendedor não encontrado"));
+        VendedorEntity vendedorEntity =
+                vendedorRepository.findByIdVendedorAndUsuarioId(
+                        idVendedor,
+                        usuarioEntity.getId()
+                ).orElseThrow(() ->
+                        new RuntimeException("Vendedor não encontrado"));
+
+        boolean possuiNotas =
+                !notaFiscalRepository
+                        .findByVendedorIdVendedor(idVendedor)
+                        .isEmpty();
+
+        if (possuiNotas) {
+            throw new RuntimeException(
+                    "Não é possível excluir este vendedor porque existem notas fiscais vinculadas a ele."
+            );
+        }
 
         vendedorRepository.delete(vendedorEntity);
     }
@@ -76,9 +96,24 @@ public class VendedorService {
         String email = jwtUtil.extrairEmailToken(token.substring(7));
         UsuarioEntity usuarioEntity = usuarioRepository.findByEmail(email)
                 .orElseThrow(()->new RuntimeException("Email não encontrado "+email));
+        validarComissao(usuarioEntity, comissao);
         VendedorEntity vendedorEntity = vendedorRepository.findByIdVendedorAndUsuarioId(idVendedor,usuarioEntity.getId())
                 .orElseThrow(()->new RuntimeException("Id não encontrado "+idVendedor));
          vendedorEntity.setComissao(comissao);
         return vendedorConverter.paraVendedorDTOResponse(vendedorRepository.save(vendedorEntity));
+    }
+    private void validarComissao(UsuarioEntity usuario, Double comissaoVendedor) {
+
+        if (comissaoVendedor == null || comissaoVendedor < 0) {
+            throw new RuntimeException("Informe uma comissão válida");
+        }
+
+        if (usuario.getComissaoTotal() == null) {
+            throw new RuntimeException("Configure primeiro a comissão total da empresa");
+        }
+
+        if (comissaoVendedor > usuario.getComissaoTotal()) {
+            throw new RuntimeException("A comissão do vendedor não pode ser maior que a comissão total da empresa");
+        }
     }
 }
